Reworked nsw_add_c_test() in CMake
With the changed structure in the directory layout of /test/, this command was changed as well.
The command now supports:
- Additional source files
- Copying the directory /test/<name>.d/resources into the build
 directory
- Disabling copying that directory.
- Installing the resources directory.

Additionally, an issue with the command that was working directory related is
now fixed.

nsw_c_test_to_cpp() was temporarily diabled.

nsw_add_cpp_test() has not been updated yet.